smartClassroom/requirements.txt
Django>=5.2
djangorestframework>=3.14
django-cors-headers>=4.3
textblob>=0.17.1
nltk>=3.8
python-dotenv>=1.0
